Address review: extend existing MockEngine to IBuildEngine4 instead of new MockEngine4
Per review feedback, the existing MockEngine in TestUtilities now implements
IBuildEngine4 (adding the IBuildEngine2/3/4 members and an in-memory
registered-task-object store) rather than introducing a separate MockEngine4
test double. IBuildEngine4 derives from IBuildEngine, so all existing
MockEngine usages remain source-compatible.

- Extend MockEngine : IBuildEngine -> IBuildEngine4
- Delete MockEngine4.cs
- Update LocateRepositoryTests to use MockEngine
